+++ src/formats/gvox_u32.cpp
@@ -27,19 +27,19 @@ GVoxU32Context::GVoxU32Context() = default;
 
 auto GVoxU32Context::create_payload(GVoxScene scene) -> GVoxPayload {
     GVoxPayload result = {};
-    result.size += sizeof(size_t);
-    for (size_t node_i = 0; node_i < scene.node_n; ++node_i) {
+    result.size += sizeof(uint64_t);
+    for (uint64_t node_i = 0; node_i < scene.node_n; ++node_i) {
         if (scene.nodes[node_i].voxels == nullptr) {
             continue;
         }
-        result.size += sizeof(size_t) * 3;
+        result.size += sizeof(uint64_t) * 3;
         result.size += scene.nodes[node_i].size_x * scene.nodes[node_i].size_y * scene.nodes[node_i].size_z * sizeof(uint32_t);
     }
-    result.data = new uint8_t[result.size];
+    result.data = new uint8_t[static_cast<size_t>(result.size)];
     auto *buffer_ptr = (uint8_t *)result.data;
     std::memcpy(buffer_ptr, &scene.node_n, sizeof(scene.node_n));
     buffer_ptr += sizeof(scene.node_n);
-    for (size_t node_i = 0; node_i < scene.node_n; ++node_i) {
+    for (uint64_t node_i = 0; node_i < scene.node_n; ++node_i) {
         if (scene.nodes[node_i].voxels == nullptr) {
             continue;
         }
@@ -49,9 +49,9 @@ auto GVoxU32Context::create_payload(GVoxScene scene) -> GVoxPayload {
         buffer_ptr += sizeof(scene.nodes[node_i].size_y);
         std::memcpy(buffer_ptr, &scene.nodes[node_i].size_z, sizeof(scene.nodes[node_i].size_z));
         buffer_ptr += sizeof(scene.nodes[node_i].size_z);
-        size_t const voxels_n = scene.nodes[node_i].size_x * scene.nodes[node_i].size_y * scene.nodes[node_i].size_z;
-        size_t const voxels_size = voxels_n * sizeof(uint32_t);
-        for (size_t voxel_i = 0; voxel_i < voxels_n; ++voxel_i) {
+        uint64_t const voxels_n = scene.nodes[node_i].size_x * scene.nodes[node_i].size_y * scene.nodes[node_i].size_z;
+        uint64_t const voxels_size = voxels_n * sizeof(uint32_t);
+        for (uint64_t voxel_i = 0; voxel_i < voxels_n; ++voxel_i) {
             auto const &i_vox = scene.nodes[node_i].voxels[voxel_i];
             uint32_t u32_voxel = 0;
             uint32_t const r = static_cast<uint32_t>(std::max(std::min(i_vox.color.x, 1.0f), 0.0f) * 255.0f);
@@ -77,21 +77,21 @@ auto GVoxU32Context::parse_payload(GVoxPayload payload) -> GVoxScene {
     GVoxScene result = {};
     auto *buffer_ptr = (uint8_t *)payload.data;
     uint8_t *buffer_sentinel = (uint8_t *)payload.data + payload.size;
-    result.node_n = *(size_t *)buffer_ptr;
+    result.node_n = *(uint64_t *)buffer_ptr;
     buffer_ptr += sizeof(result.node_n);
-    result.nodes = (GVoxSceneNode *)std::malloc(sizeof(GVoxSceneNode) * result.node_n);
-    size_t node_i = 0;
+    result.nodes = (GVoxSceneNode *)std::malloc(sizeof(GVoxSceneNode) * static_cast<size_t>(result.node_n));
+    uint64_t node_i = 0;
     while (buffer_ptr < buffer_sentinel) {
-        result.nodes[node_i].size_x = *(size_t *)buffer_ptr;
+        result.nodes[node_i].size_x = *(uint64_t *)buffer_ptr;
         buffer_ptr += sizeof(result.nodes[node_i].size_x);
-        result.nodes[node_i].size_y = *(size_t *)buffer_ptr;
+        result.nodes[node_i].size_y = *(uint64_t *)buffer_ptr;
         buffer_ptr += sizeof(result.nodes[node_i].size_y);
-        result.nodes[node_i].size_z = *(size_t *)buffer_ptr;
+        result.nodes[node_i].size_z = *(uint64_t *)buffer_ptr;
         buffer_ptr += sizeof(result.nodes[node_i].size_z);
-        size_t const voxels_n = result.nodes[node_i].size_x * result.nodes[node_i].size_y * result.nodes[node_i].size_z;
-        size_t const voxels_size = voxels_n * sizeof(GVoxVoxel);
-        result.nodes[node_i].voxels = (GVoxVoxel *)std::malloc(voxels_size);
-        for (size_t voxel_i = 0; voxel_i < voxels_n; ++voxel_i) {
+        uint64_t const voxels_n = result.nodes[node_i].size_x * result.nodes[node_i].size_y * result.nodes[node_i].size_z;
+        uint64_t const voxels_size = voxels_n * sizeof(GVoxVoxel);
+        result.nodes[node_i].voxels = (GVoxVoxel *)std::malloc(static_cast<size_t>(voxels_size));
+        for (uint64_t voxel_i = 0; voxel_i < voxels_n; ++voxel_i) {
             uint32_t u32_voxel = 0;
             std::memcpy(&u32_voxel, buffer_ptr + voxel_i * sizeof(uint32_t), sizeof(uint32_t));
             float r = static_cast<float>((u32_voxel >> 0x00) & 0xff) / 255.0f;
